backend/requirements.txt
# requirements.txt — 完全鎖定套件版本，請勿手動修改，當環境需要重現時直接 pip install -r requirements.txt

annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.115.14
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.35.0
